requirements.txt
absl-py==1.2.0
aiohttp==3.8.3
aiosignal==1.2.0
antlr4-python3-runtime==4.8
async-timeout==4.0.2
asynctest==0.13.0
attrs==22.1.0
beautifulsoup4==4.11.1
blobfile==2.0.0
cachetools==5.2.0
cffi==1.15.1
charset-normalizer==2.1.1
cycler==0.11.0
cynetworkx
Cython==0.29.32
decorator==4.4.2
depthconv-ext==0.1.0
einops==0.3.0
filelock==3.8.0
fonttools==4.37.4
freetype-py==2.3.0
frozenlist==1.3.1
fsspec==2022.8.2
ftfy==6.1.1
future==0.18.2
fvcore
gdown==4.5.3
google-auth==2.12.0
google-auth-oauthlib==0.4.6
grpcio==1.49.1
h5py==3.7.0
huggingface-hub==0.13.4
idna==3.4
imageio==2.22.1
imageio-ffmpeg==0.4.7
importlib-metadata==5.0.0
imutils==0.5.4
iopath
kiwisolver==1.4.4
lpips==0.1.4
lxml==4.9.1
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.3
mkl-fft==1.3.1
mkl-random
mkl-service==2.4.0
moviepy==1.0.2
mpi4py
multidict==6.0.2
networkx==2.3
numpy==1.21.6
oauthlib==3.2.1
olefile
omegaconf==2.1.1
opencv-python==4.2.0.32
packaging==21.3
pandas==1.3.5
Pillow==9.3.0
plyfile==0.7.4
portalocker
proglog==0.1.10
protobuf==3.19.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybind11==2.10.0
pycparser==2.21
pycryptodomex==3.15.0
pyDeprecate==0.3.1
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
pytorch-lightning==1.4.2
pytorch3d==0.7.0
pytz==2022.6
PyWavelets==1.3.0
PyYAML==6.0
regex==2022.9.13
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.19.3
scipy==1.7.3
six
soupsieve==2.3.2.post1
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
termcolor==1.1.0
test-tube==0.7.5
tifffile==2021.11.2
timm==0.6.13
torch==1.8.0
torchaudio==0.8.0a0+a751e1d
torchmetrics==0.5.0
torchvision==0.9.0
tqdm
transforms3d==0.3.1
urllib3==1.26.12
vispy==0.6.4
wcwidth==0.2.5
Werkzeug==2.2.2
yarl==1.8.1
zipp==3.8.1
git+https://github.com/CompVis/taming-transformers.git@master#egg=taming-transformers
git+https://github.com/openai/CLIP.git@main#egg=clip